This is a ipynb file which has a code of whole project.
Let’s see how to generate QR code in Python using pyqrcode module.
The module automates most of the building process for creating QR codes. This module attempts to follow the QR code standard as closely as possible. The terminology and the encoding used in pyqrcode come directly from the standard.